Various updates: - Converts views to use ORM access to the dbs - Updates the db_conn module to use Django accessor methods to get a dbs connection - Changes approach with LinkedIn API; rather than using pure REST server-side, it makes more sense to use the .js API client side for initial auth (+ potentially to import user details) and then push the secure cookie generated by that server side to get OAuth tokens.
Notes:
- We should really talk about the current dbs schema. It's set up in a
way that's not super optimal now, e.g. heavy use of varchar(3000),
including for self-joins, over-normalisation in some places etc. It's
also not possible for me to create that schema locally, as some of the
row lenghths are > 65K limits due to the very large varchar fields
involved.
- Due to the above issue, I can't test my dbs / models updates
locally. It's pretty standard code, so it should be fine, but do be
aware that there might be an odd syntax issue in there somewhere.

Change-Id: Ic0f9950fb13606d93a138220683fe5649764aeac