Re-factors the models file and dbs schema. This is a change for discussion; once we're agreed on the approach to take, I'll make any other changes required throughout the code to use the model structure and then merge to master.
Items marked DISCUSS are, somewhat self-evidently, points where it would be helpful to get your feedback and input. It's *very* likely that we'll want to add some redundancy to this schema to make querying simpler, but I'd like to get to a clean, bare-bones schema and build up from that based on learnings from developing the app / tuning queries first.

Changes:
1. Moves all choices into a hunchworks_enums.py file, and makes most of them IntegerFields.
2. Replaces most uses of varchar(1000) fields; by and large these are replaced with M2M tables. Chris - I'm happy to walk you through why long varchar fields in the way you were using them are a bad idea vs. use of relational semantics if that would be helpful?
3. Consolidates various fields, e.g. languages, skill and expertise.
4. Makes use of various Django special fields, e.g. URLField etc.
5. Adds in stub i18n code to settings
6. Adds docstrings throughout.
7. Adds a whole lot of new tables, e.g. Role, Skill etc., to represent additional logical objects in the application.

N.B. This change assumes that we're going to be using an RDBMS for application hosting; hence the heavy use of relational semantics etc.