Renamed all table names to hw_all_lowercase style after finding a standard for it. Also renamed foreign keys as well.
Created ManytoMany fields in the models.py for the classes I think needed them.
Added in a Text label M2M on the all tables diagram in the .mwb file, and gave more information about the connection in the Hunch and User diagrams.

Tested the code, everything runs as is, except the hunchworks_enums file, I am not sure why it is throwing an error on that, as I do not understand how the enums work in python. Will look into it tomorrow or some other time.

updated sql file with new database.