Edit Hunch and Create Hunch now properly pull and display tags, languages, skills, and users.
Modified edit hunch to pull the tags for the specifc hunch and display them, and displays all the users that are working on that specific hunch.
Created a Json view to display all the users for that specific hunch.
Alphabetized the json views, but kept it broken down for users and hunches.
Fixed a small error in urls file.