Fixed Signup, edit Hunch pages
Edit Hunch page now takes the tags, skills, languages, and users on the page and stores them in the DB the same way create Hunch does.

Signup page now uses the Jquery plugin to add skills and languages instead of typing it in by hand. Also saves to the database upon signup as well.