Added UserProfileWidget
Created a widget for adding users to the hunch or groups, however it is not completely finished. The userId is currently hardcoded and Adam and I need to figure out how to get the UserId or HunchId to into the data-search url.

Also fixed an error on the create a Hunch button on Profile page.