View a Hunch page now more resembles the hunch
Fixed the Tags
Fixed the Strength (still only shows positive numbers)
Shows the current evidence attached to it, but shows it all as tweets, which will need to be fixed in the future.
Displays a list of all the users working on the hunch, with links to their profile pages. In the future this will need to be changed to limit the list.