Extract embed stuff into the `djembedly` package.
You can probably see where this is going.

I'm trying not to reinvent the wheel here, but `django-embedly` is not
sufficient (for me). The point of this package is to make it easy to
convert links into embeds, while allowing us to easily add our own
embed targets. It's like, 50 lines of code.