Extract hunch permission editing to separate form.
Hunch permissions are way too complicated, so I'm extracting them to a
separate form which hopefully nobody will ever have to look at.

It should be pretty rare that permissions need to be revoked (although
obviously it should be possible). Mostly, groups and users should be
added by inviting them.